Created ScreenState sealed class for better loading UX
Now ViewModels can return a sealed class that represents if data has
loaded or not and potentially an error in the future.

Progress Bars have been added to the root content so it can be shown on
any screen while the data is loading.

Some UX optimisation was done to help clean up some of the layouts.